Everything that I had before should be state aware now. :) Fir_test now has a few test cases, and more will be needed to make sure everything is working properly.
The complex1 test case blows up horribly right now.
It evaluates perfectly well, but takes an amazingly
large amount of resources.  The .prlb file (not included) for that
one proof is 8.4 MB and MetaPRL ~64MB of RAM and 2 minutes on a
1 GHz Pentium III system.
After the proof, the "counts" were [2, ~34000] for that proof
(took me 2 steps).  After restarting MetaPRL and reloading the proof,
the counts were [2,290483].  Needless to say, something needs
to be done concerning efficiency.  As far as I can tell, things
are not being reduced in an optimal order. I'll be looking more
into this.