Documentation build system and references cleanup:
- Migrated from metaprl.bib to rc.bib. rc.bib is currently maintained by several
people (including Bob's secretary) and is likely to be of better quality (eventually?).

- Moved the editor/ml/tex directory to a more appropriate doc/latex/theories/ocaml_doc
This move is paritally done server-side, so it is *not* versioned!

- We used to have two different versions of metaprl.bib and metaprl.tex, now all TeX
inputs are in doc/latex/inputs

- Merged the book compilation scripts with the Make build subsystem.
Now the book can be built by running "make ocaml-book" from the top level.

- "THEORIES=ALL" will now include ocaml_doc as well.

- Ocaml book will be built with CM fonts by default. If you have Licida-Bright fonts,
uncomment the corresponding line in doc/latex/theories/book2.tex