- This fixes all-theories.pdf build (bug 57)
- In bytecode mode, omake should use mp.top, not mp.opt

- Made ocamldep -prl mode a bit smarter:
a) When a PRL .mli file contains "mptop", then .cmi dependes on Shell_sig and Mptop
modules
b) When a PRL file contains "interactive", "interactive_rw" or "derived", then it
depeneds on the Shell module
c) All PRL .cmx and .cmo files depend on Shell_sig and Mptop signatures (strictly speaking,
this is only true when the corresponding .mli contains an "mptop" directive), but this is an
approximation

- Printexn.catch is a bad idea now (the runtime now is as good at printing uncaught
exceptions and Printexn.catch blocks exception backtracing), got rid of all of them.

- Print a better error message when trying to cd into a theory that does not exist.

- The all-theories theory list was missing the base_theory file, and the fol_theory one
should not be there.

- Minor changes in module (AKA theory, AKA package) management in shell.

- Made "omake clean" in editor/ml a bit cleaner