This implements bug 188:
Now in prim rules syntax the "= term" extract specification is optional.
By default (e.g. if the extract specification is omited), the extract
is taken to be "default_extract{}". The system will still complain if the
default_extract opname is not declared, making the extracts truly optional
only in theories that declare such an opname.

For now, the only theory that uses this is the fir one (which I used to test
the implemention).