This commit changes the way the AutoMustComplete flag for intro annotations is implemented.
Instead of using it to check the item found by the term table _after_ it is
found, the AutoMustComplete flag is now checked _during_ the term table lookup.
This means that the AutoMustComplete entries will now be completely ignored
during the "normal" phase of autoT, allowing it to fall back to less specific
entries. The manual entires to intro now have to include an extra boolean field
--- setting it to true has the same effect as AutoMustComplete.

P.S. The meaning and implementation of CondMustComplete did not change.