In print_theory, use LsAll (as opposed to the "print nothing" empty set!).
Note, that before the recent ls options management changes we used a custom
selector that took care of automatically ingoring things like ML code. Now
notghing is ignored automatically (a good thing IMHO), which means that we
have to be more careful in placing "docoff" in all the right places.