Oops, I messed up the normal term parsing.
We have a slight problem with operator names.  We would like
to write the following.

   "operator"[(foo{x; y}):op]

However, this won't work because foo{x; y} parses
as xterm{[foo]; []; [x; y]}, so we get the following
term instead!)

   << operator[xterm{0; 0; 0}] >>

What we have to do instead is to use an iform to place
the term in the right context, so we have to use an
indirect production instead.  I've chosen $, for the
following.

   $lambda{x. x}     == operator[(lambda{x. x}):op]