Added initial quoting of rules--this currently only works for sovars/cvars without any arguments.  Arguments need more work, I'll fix it next (it is hard).  See Pmn_core_judgments for the current trivial example.
For each reflected logic, we add the rules as type *checking* predicates
on steps in a derivation.  A derivation is valid iff all the steps check
and all the terms in it are well-formed.  A proof is the least-fixed-point
of the derivations (using srec).  A term is provable if it is the final
term in a derivation.  srec gives us proof induction.