Added the structural rules for the meta-logic, including cut and thinning.
Unfortunately, these rules are ML rules, but only because we can't
express the extract properly.

Yes I know, meta-logical power is like a siren, and we can become
entralled by her beauty--to our own peril.

Still it would be cool if we could write down this stuff the right
way.  For example, thinning.

   <H>; <J> --> E
   -----------------
   <H>; A; <J> --> E

Here, <H> and <J> correspond to lists of assumptions.
Crazy, yes.  At least they are not dependent:/