Sorry----
For the short term, we need to remove the "denormalizing" rewrites from
reduceC.  This includes the following rewrites in Itt_hoas_debruijn:
   reduce_mk_bterm_base
   reduce_mk_bterm_step
   reduce_mk_bterm_empty
All of these are "denormalization" rules, because they hoist binds
out of bterms.

This is only a half-hearted attempt.  Itt_hoas_lang fails miserably
if these are removed from reduceC.