Separate the theory configuration from the build specification.
Now, each theory includes a required file called MetaprlInfo (we
can change the name, although I prefer the generic name Theory or
TheoryInfo:).

    MetaprlInfo specifies:
        THEORYNAME = ...
        THEORYDESCR = ...
        THEORY_DEPS[] = ...
        MLZFILES[] = ...
        MPFILES[] = ...
        PRINT_THEORIES[] = ...

If this is all you have (the normal case), you do not need to define an
OMakefile.  The build will proceed according to the normal rules.

If you have special concerns, you may also write an OMakefile for
the directory.  In this case, the default actions are *not* taken,
and you should specify them by hand.

Next up: configure the install directory, and allow a search
path for locating theories.