Split the judgments out of Pmn_core_terms.
An initial thought here is to define the meta-type judgment
manually for sequents, rather than use the existential types.

According to our plan, we don't plan to have wf judgments
for sequents.  However, we see that Filter_{parse,reflection}
are doing the wrong thing on actual rules (as opposed to
declare statements), so I'll fix that.